frontend: Refactor prepareSidebarItems into a hook
