+++ frontend/src/components/Sidebar/Sidebar.tsx
@@ -10,15 +10,14 @@ import { useTranslation } from 'react-i18next';
 import { useDispatch } from 'react-redux';
 import { useHistory } from 'react-router-dom';
 import helpers from '../../helpers';
-import { useCluster } from '../../lib/k8s';
 import { createRouteURL } from '../../lib/router';
 import { useTypedSelector } from '../../redux/reducers/reducers';
 import { ActionButton } from '../common';
 import CreateButton from '../common/Resource/CreateButton';
 import NavigationTabs from './NavigationTabs';
-import prepareRoutes from './prepareRoutes';
 import SidebarItem, { SidebarItemProps } from './SidebarItem';
 import { DefaultSidebars, setSidebarSelected, setWhetherSidebarOpen } from './sidebarSlice';
+import { useSidebarItems } from './useSidebarItems';
 import VersionButton from './VersionButton';
 
 export const drawerWidth = 240;
@@ -177,8 +176,6 @@ function updateItemSelected(
 }
 
 export default function Sidebar() {
-  const { t, i18n } = useTranslation(['glossary', 'translation']);
-
   const sidebar = useTypedSelector(state => state.sidebar);
   const {
     isOpen,
@@ -188,24 +185,10 @@ export default function Sidebar() {
     isTemporary: isTemporaryDrawer,
   } = useSidebarInfo();
   const isNarrowOnly = isNarrow && !canExpand;
-  const arePluginsLoaded = useTypedSelector(state => state.plugins.loaded);
   const namespaces = useTypedSelector(state => state.filter.namespaces);
   const dispatch = useDispatch();
-  const cluster = useCluster();
-  const items = React.useMemo(() => {
-    // If the sidebar is null, then it means it should not be visible.
-    if (sidebar.selected.sidebar === null) {
-      return [];
-    }
-    return prepareRoutes(t, sidebar.selected.sidebar || '');
-  }, [
-    cluster,
-    sidebar.selected.sidebar,
-    sidebar.entries,
-    sidebar.filters,
-    i18n.language,
-    arePluginsLoaded,
-  ]);
+
+  const items = useSidebarItems(sidebar?.selected?.sidebar ?? undefined);
 
   const search = namespaces.size !== 0 ? `?namespace=${[...namespaces].join('+')}` : '';
